+++ fibre/src/domain/error.rs
