+++ fibre/src/domain/mod.rs
@@ -1,0 +1,7 @@
+//! Domain types and deterministic protocol logic.
+
+pub(crate) mod blob;
+pub(crate) mod blob_header;
+pub mod config;
+pub mod error;
+pub(crate) mod payment_promise;
