+++ fibre/src/transport/mod.rs
@@ -1,0 +1,6 @@
+//! Transport abstractions and gRPC adapters.
+
+pub(crate) mod grpc_validator_client;
+pub(crate) mod host_registry;
+pub(crate) mod proto_conv;
+pub(crate) mod validator_client;
